requirements.txt

requests
python-docx
